refactor: extract optimize run loop and clock args conversion to domain layer
Move `run_optimize_loop`, `ConvergenceReason`, `OptimizeLoopResult`, `collect_optimize_partitions`, and supporting types from `commands/optimize/run.rs` to `optimize/run_loop.rs`. This resolves the dependency inversion where domain-layer tests imported from the command layer.

Move `From<&BranchSplitArgs> for BranchPointOptimizationParams` from `clock/find_best_root/params.rs` to a `branch_split_to_params` helper in `commands/clock/run.rs`, removing the domain-to-command import.

- `commands/optimize/run.rs` retains only `run_optimize` (I/O orchestrator) and `TreetimeOptimizeParams`
- 12 test files updated to import from `optimize::run_loop` instead of `commands::optimize::run`
- `optimize/__tests__/test_args.rs` retains its `commands::optimize::args` import (validates CLI arg parsing)
