+++ packages/treetime/src/optimize/__tests__/test_convergence_conditions.rs
@@ -1,6 +1,6 @@
 #[cfg(test)]
 mod tests {
-  use crate::commands::optimize::run::{ConvergenceReason, collect_optimize_partitions, run_optimize_loop};
+  use crate::optimize::run_loop::{ConvergenceReason, collect_optimize_partitions, run_optimize_loop};
   use crate::optimize::__tests__::test_convergence::test_convergence_support::tests::{
     TREE_NEWICK, setup_partitions, simple_alignment,
   };
