+++ packages/treetime/src/optimize/__tests__/test_convergence_sc2.rs
@@ -3,7 +3,7 @@ mod tests {
   use crate::alphabet::alphabet::Alphabet;
 
   use crate::ancestral::marginal::update_marginal;
-  use crate::commands::optimize::run::{collect_optimize_partitions, run_optimize_loop};
+  use crate::optimize::run_loop::{collect_optimize_partitions, run_optimize_loop};
   use crate::gtr::get_gtr::{JC69Params, jc69};
   use crate::optimize::args::BranchOptMethod;
   use crate::optimize::optimize_unified::initial_guess_mixed;
