+++ packages/treetime/src/optimize/mod.rs
@@ -10,3 +10,4 @@ pub mod optimize_eval;
 pub mod optimize_indel;
 pub mod optimize_sparse_eval;
 pub mod optimize_unified;
+pub mod run_loop;
